Export Task constructor prop types
This was an oversight from v3 which dissallowed using the `Domain`
interface correctly.

Change-type: patch
