+++ lib/task/index.ts
@@ -1,4 +1,5 @@
 export * from './tasks';
+export * from './props';
 export * from './instructions';
 export * from './helpers';
 export * from './context';
